+++ src/publicodes-state/hooks/useCurrentSimulation/useCurrentSimulation.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import useUser from '../useUser'
+import useUser from '../useUser/useUser'
 
 /**
  * A hook to get the current simulation and update it (via simulation.update() or updateCurrentSimulation())
