+++ src/publicodes-state/hooks/useRule/useRule.ts
@@ -7,15 +7,15 @@ import { utils } from 'publicodes'
 import { useContext, useMemo } from 'react'
 import { SimulationContext } from '../../providers/simulationProvider/context'
 import type { Metric } from '../../types'
-import useCurrentSimulation from '../useCurrentSimulation'
-import useChoices from './useChoices'
-import useContent from './useContent'
-import useMissing from './useMissing'
-import useNotifications from './useNotifications'
-import useQuestionsOfMosaic from './useQuestionsOfMosaic'
-import useSetValue from './useSetValue'
-import useType from './useType'
-import useValue from './useValue'
+import useCurrentSimulation from '../useCurrentSimulation/useCurrentSimulation'
+import useChoices from './hooks/useChoices'
+import useContent from './hooks/useContent'
+import useMissing from './hooks/useMissing'
+import useNotifications from './hooks/useNotifications'
+import useQuestionsOfMosaic from './hooks/useQuestionsOfMosaic'
+import useSetValue from './hooks/useSetValue'
+import useType from './hooks/useType'
+import useValue from './hooks/useValue'
 
 /**
  * A hook to get and set every information about a specific rule
