+++ src/publicodes-state/hooks/useUser/hooks/useNorthStar.ts
@@ -1,6 +1,6 @@
 import type { NorthStarType, NorthStarValue } from '@/types/northstar'
 import type { Dispatch, SetStateAction } from 'react'
-import type { User } from '../../types'
+import type { User } from '../../../types'
 
 type Props = {
   setUser: Dispatch<SetStateAction<User>>
