+++ src/publicodes-state/hooks/useUser/hooks/useSimulations.ts
@@ -4,7 +4,7 @@ import { generateSimulation } from '@/helpers/simulation/generateSimulation'
 import type { Migration } from '@publicodes/tools/migration'
 import type { Dispatch, SetStateAction } from 'react'
 import { useCallback, useMemo } from 'react'
-import type { Simulation, UpdateCurrentSimulationProps } from '../../types'
+import type { Simulation, UpdateCurrentSimulationProps } from '../../../types'
 
 type Props = {
   simulations: Simulation[]
