+++ src/publicodes-state/index.ts
@@ -1,13 +1,13 @@
-export { default as FormProvider } from './providers/formProvider'
-export { default as SimulationProvider } from './providers/simulationProvider'
-export { default as UserProvider } from './providers/userProvider'
+export { default as FormProvider } from './providers/formProvider/provider'
+export { default as SimulationProvider } from './providers/simulationProvider/provider'
+export { default as UserProvider } from './providers/userProvider/provider'
 
-export { default as useActions } from './hooks/useActions'
-export { default as useCurrentSimulation } from './hooks/useCurrentSimulation'
-export { default as useDisposableEngine } from './hooks/useDisposableEngine'
-export { default as useEngine } from './hooks/useEngine'
-export { default as useForm } from './hooks/useForm'
-export { default as useRule } from './hooks/useRule'
-export { default as useSimulation } from './hooks/useSimulation'
-export { default as useTempEngine } from './hooks/useTempEngine'
-export { default as useUser } from './hooks/useUser'
+export { default as useActions } from './hooks/useActions/useActions'
+export { default as useCurrentSimulation } from './hooks/useCurrentSimulation/useCurrentSimulation'
+export { default as useDisposableEngine } from './hooks/useDisposableEngine/useDisposableEngine'
+export { default as useEngine } from './hooks/useEngine/useEngine'
+export { default as useForm } from './hooks/useForm/useForm'
+export { default as useRule } from './hooks/useRule/useRule'
+export { default as useSimulation } from './hooks/useSimulation/useSimulation'
+export { default as useTempEngine } from './hooks/useTempEngine/useTempEngine'
+export { default as useUser } from './hooks/useUser/useUser'
