+++ src/publicodes-state/providers/formProvider/Provider.tsx
@@ -4,21 +4,18 @@ import { useCurrentSimulation } from '@/publicodes-state'
 
 import type { DottedName } from '@incubateur-ademe/nosgestesclimat'
 import type { PropsWithChildren } from 'react'
-import { useContext } from 'react'
+import { useContext, useMemo } from 'react'
 import { SimulationContext } from '../simulationProvider/context'
 import FormContext from './context'
-import useCurrent from './useCurrent'
-import useProgression from './useProgression'
-import useQuestions from './useQuestions'
+import useCurrent from './hooks/useCurrent'
+import useProgression from './hooks/useProgression'
+import useQuestions from './hooks/useQuestions'
 
 type Props = {
   root?: DottedName
 }
 
-export default function FormProvider({
-  root = 'bilan',
-  children,
-}: PropsWithChildren<Props>) {
+function FormProvider({ root = 'bilan', children }: PropsWithChildren<Props>) {
   const {
     categories,
     subcategories,
@@ -76,4 +73,28 @@ export default function FormProvider({
       {children}
     </FormContext.Provider>
   )
+}
+
+/**
+ * This is not the real provider but a failsafe: if root is invalid we do not go further
+ */
+export default function FailSafeFormProvider({
+  root = 'bilan',
+  children,
+}: PropsWithChildren<{
+  root?: DottedName
+}>) {
+  const { safeEvaluate, rules } = useContext(SimulationContext)
+
+  const isRootSafe = useMemo<boolean>(() => {
+    if (!rules) return true
+
+    return safeEvaluate(root) ? true : false
+  }, [safeEvaluate, root, rules])
+
+  if (!isRootSafe) {
+    window.location = '/404' as unknown as Location
+    return
+  }
+  return <FormProvider root={root}>{children}</FormProvider>
 }
