+++ src/publicodes-state/providers/simulationProvider/hooks/useCategories.ts
