+++ src/publicodes-state/providers/userProvider/hooks/usePersistentSimulations.ts
@@ -2,7 +2,7 @@ import { generateSimulation } from '@/helpers/simulation/generateSimulation'
 import { getIsLocalStorageAvailable } from '@/utils/getIsLocalStorageAvailable'
 import type { Migration } from '@publicodes/tools/migration'
 import { useEffect, useState } from 'react'
-import type { Simulation } from '../../types'
+import type { Simulation } from '../../../types'
 
 const isLocalStorageAvailable = getIsLocalStorageAvailable()
