+++ src/publicodes-state/providers/userProvider/hooks/usePersistentUser.ts
@@ -1,7 +1,7 @@
 import { getIsLocalStorageAvailable } from '@/utils/getIsLocalStorageAvailable'
 import { useEffect, useState } from 'react'
 import { v4 as uuid } from 'uuid'
-import type { RegionFromGeolocation, User } from '../../types'
+import type { RegionFromGeolocation, User } from '../../../types'
 
 const isLocalStorageAvailable = getIsLocalStorageAvailable()
